link components to button
